requirements.txt
python-dotenv
psycopg2-binary
sqlalchemy
alembic
FastAPI[all]
uvicorn
pydantic